Move types out of top-level Internal namespaces
To correspond to the interfaces which they implement, etc.
