+++ src/EFCore.InMemory/Infrastructure/Internal/InMemoryModelValidator.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Utilities;
 
-namespace Microsoft.EntityFrameworkCore.InMemory.Internal
+namespace Microsoft.EntityFrameworkCore.InMemory.Infrastructure.Internal
 {
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
