+++ src/EFCore.SqlServer/Extensions/Internal/SqlServerLoggerExtensions.cs
@@ -4,9 +4,10 @@
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.SqlServer.Internal;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.EntityFrameworkCore.SqlServer.Internal
+namespace Microsoft.EntityFrameworkCore.SqlServer.Extensions.Internal
 {
     /// <summary>
     ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
