+++ src/EFCore.SqlServer/Infrastructure/Internal/SqlServerModelValidator.cs
@@ -10,10 +10,12 @@ using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using Microsoft.EntityFrameworkCore.SqlServer.Extensions.Internal;
+using Microsoft.EntityFrameworkCore.SqlServer.Internal;
 using Microsoft.EntityFrameworkCore.SqlServer.Metadata.Internal;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.EntityFrameworkCore.SqlServer.Internal
+namespace Microsoft.EntityFrameworkCore.SqlServer.Infrastructure.Internal
 {
     /// <summary>
     ///     <para>
