+++ src/EFCore.Sqlite.Core/Extensions/Internal/SqliteLoggerExtensions.cs
