+++ src/EFCore.Sqlite.Core/Infrastructure/Internal/SqliteModelValidator.cs
@@ -2,12 +2,15 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Linq;
+using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Utilities;
+using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Sqlite.Internal;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.EntityFrameworkCore.Internal
+namespace Microsoft.EntityFrameworkCore.Sqlite.Infrastructure.Internal
 {
     /// <summary>
     ///     <para>
@@ -22,7 +25,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
     ///         This service cannot depend on services registered as <see cref="ServiceLifetime.Scoped" />.
     ///     </para>
     /// </summary>
-    public class LoggingOptions : ILoggingOptions
+    public class SqliteModelValidator : RelationalModelValidator
     {
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -30,12 +33,11 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual void Initialize(IDbContextOptions options)
+        public SqliteModelValidator(
+            [NotNull] ModelValidatorDependencies dependencies,
+            [NotNull] RelationalModelValidatorDependencies relationalDependencies)
+            : base(dependencies, relationalDependencies)
         {
-            var coreOptions = options.FindExtension<CoreOptionsExtension>() ?? new CoreOptionsExtension();
-
-            IsSensitiveDataLoggingEnabled = coreOptions.IsSensitiveDataLoggingEnabled;
-            WarningsConfiguration = coreOptions.WarningsConfiguration;
         }
 
         /// <summary>
@@ -44,28 +46,27 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual void Validate(IDbContextOptions options)
+        public override void Validate(IModel model, IDiagnosticsLogger<DbLoggerCategory.Model.Validation> logger)
         {
-            var coreOptions = options.FindExtension<CoreOptionsExtension>() ?? new CoreOptionsExtension();
+            base.Validate(model, logger);
 
-            if (IsSensitiveDataLoggingEnabled != coreOptions.IsSensitiveDataLoggingEnabled)
+            ValidateNoSchemas(model, logger);
+            ValidateNoSequences(model, logger);
+        }
+
+        /// <summary>
+        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+        ///     any release. You should only use it directly in your code with extreme caution and knowing that
+        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        /// </summary>
+        protected virtual void ValidateNoSchemas(
+            [NotNull] IModel model,
+            [NotNull] IDiagnosticsLogger<DbLoggerCategory.Model.Validation> logger)
+        {
+            foreach (var entityType in model.GetEntityTypes().Where(e => e.GetSchema() != null))
             {
-                Check.DebugAssert(coreOptions.InternalServiceProvider != null, "InternalServiceProvider is null");
-
-                throw new InvalidOperationException(
-                    CoreStrings.SingletonOptionChanged(
-                        nameof(DbContextOptionsBuilder.EnableSensitiveDataLogging),
-                        nameof(DbContextOptionsBuilder.UseInternalServiceProvider)));
-            }
-
-            if (WarningsConfiguration?.GetServiceProviderHashCode() != coreOptions.WarningsConfiguration?.GetServiceProviderHashCode())
-            {
-                Check.DebugAssert(coreOptions.InternalServiceProvider != null, "InternalServiceProvider is null");
-
-                throw new InvalidOperationException(
-                    CoreStrings.SingletonOptionChanged(
-                        nameof(DbContextOptionsBuilder.ConfigureWarnings),
-                        nameof(DbContextOptionsBuilder.UseInternalServiceProvider)));
+                logger.SchemaConfiguredWarning(entityType, entityType.GetSchema());
             }
         }
 
@@ -75,7 +76,15 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual bool IsSensitiveDataLoggingEnabled { get; private set; }
+        protected virtual void ValidateNoSequences(
+            [NotNull] IModel model,
+            [NotNull] IDiagnosticsLogger<DbLoggerCategory.Model.Validation> logger)
+        {
+            foreach (var sequence in model.GetSequences())
+            {
+                logger.SequenceConfiguredWarning(sequence);
+            }
+        }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -83,14 +92,28 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual bool IsSensitiveDataLoggingWarned { get; set; }
+        protected override void ValidateCompatible(
+            IProperty property,
+            IProperty duplicateProperty,
+            string columnName,
+            in StoreObjectIdentifier storeObject,
+            IDiagnosticsLogger<DbLoggerCategory.Model.Validation> logger)
+        {
+            base.ValidateCompatible(property, duplicateProperty, columnName, storeObject, logger);
 
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public virtual WarningsConfiguration WarningsConfiguration { get; private set; }
+            var propertySrid = property.GetSrid(storeObject);
+            var duplicatePropertySrid = duplicateProperty.GetSrid(storeObject);
+            if (propertySrid != duplicatePropertySrid)
+            {
+                throw new InvalidOperationException(
+                    SqliteStrings.DuplicateColumnNameSridMismatch(
+                        duplicateProperty.DeclaringEntityType.DisplayName(),
+                        duplicateProperty.Name,
+                        property.DeclaringEntityType.DisplayName(),
+                        property.Name,
+                        columnName,
+                        storeObject.DisplayName()));
+            }
+        }
     }
 }
