+++ src/EFCore/Diagnostics/Internal/DiagnosticsLogger.cs
@@ -1,14 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Collections.Generic;
+using System.Diagnostics;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Diagnostics;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 
-namespace Microsoft.EntityFrameworkCore.Internal
+namespace Microsoft.EntityFrameworkCore.Diagnostics.Internal
 {
     /// <summary>
     ///     <para>
@@ -24,7 +22,8 @@ namespace Microsoft.EntityFrameworkCore.Internal
     ///         The implementation does not need to be thread-safe.
     ///     </para>
     /// </summary>
-    public class DatabaseFacadeDependencies : IDatabaseFacadeDependencies
+    public class DiagnosticsLogger<TLoggerCategory> : IDiagnosticsLogger<TLoggerCategory>
+        where TLoggerCategory : LoggerCategory<TLoggerCategory>, new()
     {
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -32,20 +31,20 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public DatabaseFacadeDependencies(
-            [NotNull] IDbContextTransactionManager transactionManager,
-            [NotNull] IDatabaseCreator databaseCreator,
-            [NotNull] IExecutionStrategyFactory executionStrategyFactory,
-            [NotNull] IEnumerable<IDatabaseProvider> databaseProviders,
-            [NotNull] IDiagnosticsLogger<DbLoggerCategory.Database.Command> commandLogger,
-            [NotNull] IConcurrencyDetector concurrencyDetector)
+        public DiagnosticsLogger(
+            [NotNull] ILoggerFactory loggerFactory,
+            [NotNull] ILoggingOptions loggingOptions,
+            [NotNull] DiagnosticSource diagnosticSource,
+            [NotNull] LoggingDefinitions loggingDefinitions,
+            [NotNull] IDbContextLogger contextLogger,
+            [CanBeNull] IInterceptors interceptors = null)
         {
-            TransactionManager = transactionManager;
-            DatabaseCreator = databaseCreator;
-            ExecutionStrategyFactory = executionStrategyFactory;
-            DatabaseProviders = databaseProviders;
-            CommandLogger = commandLogger;
-            ConcurrencyDetector = concurrencyDetector;
+            DiagnosticSource = diagnosticSource;
+            Definitions = loggingDefinitions;
+            DbContextLogger = contextLogger;
+            Logger = loggerFactory.CreateLogger(new TLoggerCategory());
+            Options = loggingOptions;
+            Interceptors = interceptors;
         }
 
         /// <summary>
@@ -54,7 +53,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IDbContextTransactionManager TransactionManager { get; }
+        public virtual ILoggingOptions Options { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -62,7 +61,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IDatabaseCreator DatabaseCreator { get; }
+        public virtual ILogger Logger { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -70,7 +69,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IExecutionStrategyFactory ExecutionStrategyFactory { get; }
+        public virtual IInterceptors Interceptors { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -78,7 +77,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IEnumerable<IDatabaseProvider> DatabaseProviders { get; }
+        public virtual DiagnosticSource DiagnosticSource { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -86,7 +85,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IDiagnosticsLogger<DbLoggerCategory.Database.Command> CommandLogger { get; }
+        public virtual LoggingDefinitions Definitions { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -94,6 +93,31 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IConcurrencyDetector ConcurrencyDetector { get; }
+        public virtual IDbContextLogger DbContextLogger { get; }
+
+        /// <summary>
+        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+        ///     any release. You should only use it directly in your code with extreme caution and knowing that
+        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        /// </summary>
+        public virtual bool ShouldLogSensitiveData()
+        {
+            var options = Options;
+            if (options == null)
+            {
+                return false;
+            }
+
+            if (options.IsSensitiveDataLoggingEnabled
+                && !options.IsSensitiveDataLoggingWarned)
+            {
+                this.SensitiveDataLoggingEnabledWarning();
+
+                options.IsSensitiveDataLoggingWarned = true;
+            }
+
+            return options.IsSensitiveDataLoggingEnabled;
+        }
     }
 }
