+++ src/EFCore/Infrastructure/Internal/CoreOptions.cs
@@ -2,13 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Threading;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Utilities;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.EntityFrameworkCore.Internal
+namespace Microsoft.EntityFrameworkCore.Infrastructure.Internal
 {
     /// <summary>
     ///     <para>
@@ -18,40 +17,24 @@ namespace Microsoft.EntityFrameworkCore.Internal
     ///         doing so can result in application failures when updating to a new Entity Framework Core release.
     ///     </para>
     ///     <para>
-    ///         The service lifetime is <see cref="ServiceLifetime.Scoped" />. This means that each
-    ///         <see cref="DbContext" /> instance will use its own instance of this service.
-    ///         The implementation may depend on other services registered with any lifetime.
-    ///         The implementation does not need to be thread-safe.
+    ///         The service lifetime is <see cref="ServiceLifetime.Singleton" />. This means a single instance
+    ///         is used by many <see cref="DbContext" /> instances. The implementation must be thread-safe.
+    ///         This service cannot depend on services registered as <see cref="ServiceLifetime.Scoped" />.
     ///     </para>
     /// </summary>
-    public class ConcurrencyDetector : IConcurrencyDetector
+    public class CoreSingletonOptions : ICoreSingletonOptions
     {
-        private int _inCriticalSection;
-        private static readonly AsyncLocal<bool> _threadHasLock = new AsyncLocal<bool>();
-        private int _refCount;
-
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
         ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual ConcurrencyDetectorCriticalSectionDisposer EnterCriticalSection()
+        public virtual void Initialize(IDbContextOptions options)
         {
-            if (Interlocked.CompareExchange(ref _inCriticalSection, 1, 0) == 1)
-            {
-                if (!_threadHasLock.Value)
-                {
-                    throw new InvalidOperationException(CoreStrings.ConcurrentMethodInvocation);
-                }
-            }
-            else
-            {
-                _threadHasLock.Value = true;
-            }
+            var coreOptions = options.FindExtension<CoreOptionsExtension>() ?? new CoreOptionsExtension();
 
-            _refCount++;
-            return new ConcurrencyDetectorCriticalSectionDisposer(this);
+            AreDetailedErrorsEnabled = coreOptions.DetailedErrorsEnabled;
         }
 
         /// <summary>
@@ -60,15 +43,27 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual void ExitCriticalSection()
+        public virtual void Validate(IDbContextOptions options)
         {
-            Check.DebugAssert(_inCriticalSection == 1, "Expected to be in a critical section");
+            var coreOptions = options.FindExtension<CoreOptionsExtension>() ?? new CoreOptionsExtension();
 
-            if (--_refCount == 0)
+            if (AreDetailedErrorsEnabled != coreOptions.DetailedErrorsEnabled)
             {
-                _threadHasLock.Value = false;
-                _inCriticalSection = 0;
+                Check.DebugAssert(coreOptions.InternalServiceProvider != null, "InternalServiceProvider is null");
+
+                throw new InvalidOperationException(
+                    CoreStrings.SingletonOptionChanged(
+                        nameof(DbContextOptionsBuilder.EnableDetailedErrors),
+                        nameof(DbContextOptionsBuilder.UseInternalServiceProvider)));
             }
         }
+
+        /// <summary>
+        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
+        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
+        ///     any release. You should only use it directly in your code with extreme caution and knowing that
+        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        /// </summary>
+        public virtual bool AreDetailedErrorsEnabled { get; private set; }
     }
 }
