+++ src/EFCore/Storage/Internal/DatabaseFacadeDependencies.cs
@@ -1,13 +1,14 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Diagnostics;
+using System.Collections.Generic;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Diagnostics;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
-namespace Microsoft.EntityFrameworkCore.Internal
+namespace Microsoft.EntityFrameworkCore.Storage.Internal
 {
     /// <summary>
     ///     <para>
@@ -23,8 +24,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
     ///         The implementation does not need to be thread-safe.
     ///     </para>
     /// </summary>
-    public class DiagnosticsLogger<TLoggerCategory> : IDiagnosticsLogger<TLoggerCategory>
-        where TLoggerCategory : LoggerCategory<TLoggerCategory>, new()
+    public class DatabaseFacadeDependencies : IDatabaseFacadeDependencies
     {
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -32,20 +32,20 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public DiagnosticsLogger(
-            [NotNull] ILoggerFactory loggerFactory,
-            [NotNull] ILoggingOptions loggingOptions,
-            [NotNull] DiagnosticSource diagnosticSource,
-            [NotNull] LoggingDefinitions loggingDefinitions,
-            [NotNull] IDbContextLogger contextLogger,
-            [CanBeNull] IInterceptors interceptors = null)
+        public DatabaseFacadeDependencies(
+            [NotNull] IDbContextTransactionManager transactionManager,
+            [NotNull] IDatabaseCreator databaseCreator,
+            [NotNull] IExecutionStrategyFactory executionStrategyFactory,
+            [NotNull] IEnumerable<IDatabaseProvider> databaseProviders,
+            [NotNull] IDiagnosticsLogger<DbLoggerCategory.Database.Command> commandLogger,
+            [NotNull] IConcurrencyDetector concurrencyDetector)
         {
-            DiagnosticSource = diagnosticSource;
-            Definitions = loggingDefinitions;
-            DbContextLogger = contextLogger;
-            Logger = loggerFactory.CreateLogger(new TLoggerCategory());
-            Options = loggingOptions;
-            Interceptors = interceptors;
+            TransactionManager = transactionManager;
+            DatabaseCreator = databaseCreator;
+            ExecutionStrategyFactory = executionStrategyFactory;
+            DatabaseProviders = databaseProviders;
+            CommandLogger = commandLogger;
+            ConcurrencyDetector = concurrencyDetector;
         }
 
         /// <summary>
@@ -54,7 +54,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual ILoggingOptions Options { get; }
+        public virtual IDbContextTransactionManager TransactionManager { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -62,7 +62,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual ILogger Logger { get; }
+        public virtual IDatabaseCreator DatabaseCreator { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -70,7 +70,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IInterceptors Interceptors { get; }
+        public virtual IExecutionStrategyFactory ExecutionStrategyFactory { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -78,7 +78,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual DiagnosticSource DiagnosticSource { get; }
+        public virtual IEnumerable<IDatabaseProvider> DatabaseProviders { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -86,7 +86,7 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual LoggingDefinitions Definitions { get; }
+        public virtual IDiagnosticsLogger<DbLoggerCategory.Database.Command> CommandLogger { get; }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
@@ -94,31 +94,6 @@ namespace Microsoft.EntityFrameworkCore.Internal
         ///     any release. You should only use it directly in your code with extreme caution and knowing that
         ///     doing so can result in application failures when updating to a new Entity Framework Core release.
         /// </summary>
-        public virtual IDbContextLogger DbContextLogger { get; }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public virtual bool ShouldLogSensitiveData()
-        {
-            var options = Options;
-            if (options == null)
-            {
-                return false;
-            }
-
-            if (options.IsSensitiveDataLoggingEnabled
-                && !options.IsSensitiveDataLoggingWarned)
-            {
-                this.SensitiveDataLoggingEnabledWarning();
-
-                options.IsSensitiveDataLoggingWarned = true;
-            }
-
-            return options.IsSensitiveDataLoggingEnabled;
-        }
+        public virtual IConcurrencyDetector ConcurrencyDetector { get; }
     }
 }
